\name{iWebPlots-package}
\alias{iWebPlots-package}
\alias{iWebPlots}
\docType{package}
\title{
Interactive web-based plots
}
\description{
The iWebPlots package provides methods for the creation of interactive web-based plots, generated directly via the R statistical environment.
Interactivity is implemented using the fundamental HTML image map technology, so no additional software such as applets or plug-ins is required.
Additional features include dynamic tooltips and text annotations as well as asynchronous alternation between two- and three-dimensional scatter plots.
}
\details{
\tabular{ll}{
Package: \tab iWebPlots\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2011-08-16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Eleni Chatzimichali <e.chatzimichali@cranfield.ac.uk>, Conrad Bessant <c.bessant@cranfield.ac.uk>

Maintainer:
Eleni Chatzimichali <e.chatzimichali@cranfield.ac.uk>
}
\references{
http://www.cranfield.ac.uk/health/researchareas/bioinformatics/page53292.html
}
